requirements.txt
streamlit==1.28.1
pandas==2.0.3
sentence-transformers==2.2.2
chromadb==0.4.10
openai==1.3.0
spacy==3.6.1
ollama==0.0.14
python-dotenv==1.0.0
deep-translator==1.11.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
scikit-learn==1.3.0
langchain==0.0.310
langchain-openai==0.0.5
langchain-experimental==0.0.51
